Basic refactor
