+++ Pipfile
@@ -1,0 +1,13 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+pykube = "*"
+pytz = "*"
+
+[dev-packages]
+
+[requires]
+python_version = "3.7"
